Refactor display task function
