+++ src/utils.js
@@ -1,3 +1,5 @@
+import { openEdit } from './task'
+
 const updatedTasks = () => {
   return JSON.parse(localStorage.getItem('tasks'));
 }
@@ -21,10 +23,73 @@ const validateInput = (id) => {
   }
 }
 
+const displayArrOfTasks = (arr) => {
+  const p = arr.map(task => (
+    `<div class='d-flex f-column taskWrapper'>
+      <div class='singleTask d-flex'>
+        <div>
+            <h3>${task.name}</h3>
+            <p>${task.date.slice(5)}</p>
+        </div>
+      </div>
+      <div class='hiddenItems'>
+        <p>${task.desc}</p>
+        <p>${task.priority}</p>
+        <p>${task.project}</p>
+      </div>
+    </div>`
+  )).join('\n');
+  document.querySelector('#tasks-container').innerHTML = p;
+  editTask();
+  expandBtn();
+  expandAction();
+}
+
+const expandBtn = () => {
+  const tasks = document.querySelectorAll('.singleTask');
+  for(let i = 0; i < tasks.length; i += 1){
+    const hidden = document.createElement('span');
+    hidden.innerText = 'show';
+    hidden.classList.add('showTaskBtn');
+    tasks[i].appendChild(hidden);
+  }
+}
+
+const expandAction = () => {
+  const tasks = document.querySelectorAll('.singleTask');
+  for (let i = 0; i < tasks.length; i++) {
+    const hiddenBtns = document.querySelectorAll('.showTaskBtn');
+    hiddenBtns[i].addEventListener('click', () => {
+      const displayItems = tasks[i].nextElementSibling;
+      if(displayItems.style.display === 'block') {
+        hiddenBtns[i].innerText = 'show'
+        displayItems.style.display = 'none';
+      } else {
+        hiddenBtns[i].innerText = 'hide'
+        displayItems.style.display = 'block'
+      }
+    })
+  }
+}
+
+const editTask = () => {
+  const tasks = document.querySelectorAll('.singleTask');
+  for(let i = 0; i < tasks.length; i += 1){
+    const editBtn = document.createElement('span');
+    editBtn.innerText = 'edit';
+    editBtn.classList.add('material-icons', 'editTaskBtn');
+    editBtn.addEventListener('click', () => {
+      openEdit(i);
+    });
+    tasks[i].appendChild(editBtn);
+  }
+}
+
 export {
   updateProjects,
   updatedProjects,
   updateTasks,
   updatedTasks,
   validateInput,
+  displayArrOfTasks,
 }
